# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Foohut Documentation Platform API
  version: 1.0.0
  description: |
    AI-native documentation platform API with hierarchical content organization,
    block-based editing, Git synchronization, and AI-powered features.

    ## Architecture
    - **Organization** > **Collection** > **Space** > **Page** > **Block**
    - JWT-based authentication with role-based access control
    - Cursor-based pagination for list endpoints
    - Rate limiting with standard headers

    ## Authentication
    All endpoints (except auth) require a valid JWT Bearer token.
    Include the token in the Authorization header: `Bearer <token>`
  contact:
    name: Foohut API Support
    email: api@foohut.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.foohut.com/v1
    description: Production server
  - url: https://staging-api.foohut.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: SecureP@ssw0rd!
              name: John Doe
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: CONFLICT
                  message: Email already registered
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Login with email and password to obtain JWT tokens
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: SecureP@ssw0rd!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: UNAUTHORIZED
                  message: Invalid email or password
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidate the current session and refresh token
      operationId: logoutUser
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /spaces/{spaceId}:
    get:
      tags:
        - Spaces
      summary: Get space
      description: Retrieve space details by ID
      operationId: getSpace
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      responses:
        '200':
          description: Space retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Spaces
      summary: Update space
      description: Update space details
      operationId: updateSpace
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSpaceRequest'
      responses:
        '200':
          description: Space updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Spaces
      summary: Delete space
      description: Permanently delete a space and all its pages
      operationId: deleteSpace
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      responses:
        '204':
          description: Space deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /spaces/{spaceId}/pages:
    get:
      tags:
        - Pages
      summary: List pages
      description: List all pages in a space (hierarchical tree structure)
      operationId: listPages
      parameters:
        - $ref: '#/components/parameters/SpaceId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: parentId
          in: query
          description: Filter by parent page ID (null for root pages)
          schema:
            type: string
            format: uuid
            nullable: true
        - name: includeContent
          in: query
          description: Include page content in response
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Pages retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Pages
      summary: Create page
      description: Create a new page in a space
      operationId: createPage
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePageRequest'
            example:
              title: Getting Started
              slug: getting-started
              parentId: null
              content:
                blocks:
                  - type: heading
                    level: 1
                    content: Getting Started
                  - type: paragraph
                    content: Welcome to our documentation!
      responses:
        '201':
          description: Page created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Page slug already exists in this space
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  parameters:
...
    SpaceId:
      name: spaceId
      in: path
      required: true
      description: Space unique identifier
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      description: Pagination cursor for next page
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: BAD_REQUEST
              message: Validation failed
              details:
                - field: email
                  message: Invalid email format
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer realm="api"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: You do not have permission to perform this action
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
    RateLimited:
      description: Too many requests - rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per time window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when the rate limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: RATE_LIMITED
              message: Rate limit exceeded. Please retry after 60 seconds.
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              description: Additional error details
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            requestId:
              type: string
              description: Request ID for support reference
...
    PaginationMeta:
      type: object
      properties:
        cursor:
          type: string
          nullable: true
          description: Cursor for the next page (null if no more pages)
        hasMore:
          type: boolean
          description: Whether more items exist
        total:
          type: integer
          description: Total count of items (when available)

    # --------------------------------------------------------------------------
    # AUTH SCHEMAS
    # --------------------------------------------------------------------------
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          description: Must contain uppercase, lowercase, number, and special character
        name:
          type: string
          minLength: 1
          maxLength: 100
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (expires in 15 minutes)
        refreshToken:
          type: string
          description: Refresh token (expires in 7 days)
        expiresIn:
          type: integer
          description: Access token expiration time in seconds
        tokenType:
          type: string
          default: Bearer
        user:
          $ref: '#/components/schemas/User'

    # --------------------------------------------------------------------------
    # USER SCHEMAS
    # --------------------------------------------------------------------------
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Space:
      type: object
      properties:
        id:
          type: string
          format: uuid
        collectionId:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
          pattern: ^[a-z0-9-]+$
        description:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [public, private, unlisted]
        emoji:
          type: string
          nullable: true
        pagesCount:
          type: integer
        gitSyncEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateSpaceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        visibility:
          type: string
          enum: [public, private, unlisted]
        emoji:
          type: string
          maxLength: 10
          nullable: true
...
    Page:
      type: object
      properties:
        id:
          type: string
          format: uuid
        spaceId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        slug:
          type: string
          pattern: ^[a-z0-9-]+$
        description:
          type: string
          nullable: true
        emoji:
          type: string
          nullable: true
        path:
          type: string
          description: Full path from root (e.g., /guides/getting-started)
        status:
          type: string
          enum: [draft, published, archived]
        visibility:
          type: string
          enum: [public, private, unlisted]
        version:
          type: integer
        publishedVersion:
          type: integer
          nullable: true
        authorId:
          type: string
          format: uuid
        lastEditorId:
          type: string
          format: uuid
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        content:
          $ref: '#/components/schemas/PageContent'
    PageList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Page'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    CreatePageRequest:
      type: object
      required:
        - title
        - slug
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          pattern: ^[a-z0-9-]+$
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        emoji:
          type: string
          maxLength: 10
        parentId:
          type: string
          format: uuid
          nullable: true
        content:
          $ref: '#/components/schemas/PageContent'
...
    PageContent:
      type: object
      properties:
        version:
          type: integer
        blocks:
          type: array
          items:
            $ref: '#/components/schemas/Block'
...
    Block:
      type: object
      required:
        - id
        - type
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/BlockType'
        content:
          oneOf:
            - type: string
            - type: object
            - type: array
          description: Block content (structure varies by type)
        properties:
          type: object
          additionalProperties: true
          description: Block-specific properties
        children:
          type: array
          items:
            $ref: '#/components/schemas/Block'
          description: Nested blocks (for list items, toggles, etc.)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BlockType:
      type: string
      enum:
        - paragraph
        - heading
        - code
        - quote
        - callout
        - list
        - listItem
        - todo
        - table
        - tableRow
        - tableCell
        - image
        - video
        - embed
        - file
        - divider
        - toggleList
        - tabs
        - tab
        - hint
        - api
        - swagger
        - openapi
        - math
        - mermaid
...